* various bug fixes and optimisations * implemented facebook post progress indicator * fixed the poi grouping by duration from the time poi was created
git-svn-id: https://blocshop.cz:8040/svn/triptracktizen/trunk@65 492e496e-58ca-f445-8ef8-0a43168f6385

+++ workspace/TripTrack/inc/dao/StorageManager.h
@@ -28,7 +28,7 @@ public:
 	Tizen::Io::DbEnumerator* PerformTransaction(
 			Tizen::Io::DbStatement* statement);
 	Tizen::Base::Collection::LinkedListT<POI*>* GetPois(void);
-	Tizen::Base::Collection::HashMapT<long long int,
+	Tizen::Base::Collection::HashMapT<int,
 			Tizen::Base::Collection::LinkedListT<POI*>*>* GetPoiHash(void);
 	Tizen::Base::Collection::LinkedListT<TTMedia*>* GetMedia(
 			long long int poiId);
@@ -40,10 +40,10 @@ public:
 	FacebookAccessToken GetFacebookCredentials(void) const;
 	void UpdateFacebookCredentials(FacebookAccessToken& credentials);
 
-
 private:
 	StorageManager();
 	static StorageManager* __pSelf;
+
 	Tizen::Io::Database* __pDb;
 	FacebookAccessToken __facebookCredentials;
 };
